Compile fix for prev. commit

+++ Aquaria/Entity.h
@@ -486,7 +486,7 @@ public:
 
 	virtual bool canSetBoneLock();
 
-	void initHair(int numSegments, int segmentLength, int width, const std::string &tex);
+	void initHair(int numSegments, float segmentLength, float width, const std::string &tex);
 	void updateHair(float dt);
 	void setHairHeadPosition(const Vector &pos);
 	void exertHairForce(const Vector &force, float dt);
